fix: 안 쓰는 코드 삭제 및 정리
