+++ src/pages/index.tsx
@@ -1,8 +1,4 @@
-import Head from "next/head";
-import Image from "next/image";
 import { Inter } from "next/font/google";
-import styles from "@/styles/Home.module.css";
-import { Box } from "@mui/material";
 import Main from "@/components/main/Main";
 
 const inter = Inter({ subsets: ["latin"] });
