--- src/pages/my/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const index = () => {
-  return (
-    <div>my</div>
-  )
-}
-
-export default index
