+++ src/pages/register/nickname/index.tsx
@@ -1,54 +1,10 @@
-import Header from '@/components/common/Header';
-import Progress from '@/components/common/Progress';
-import { useUser } from '@/hook/useUser';
-import { setColor } from '@/utils/setColor';
-import { Box, TextField } from '@mui/material';
-import { useRouter } from 'next/router';
-import React, { useState } from 'react'
+import Nickname from '@/components/register/Nickname';
+import React from 'react'
 
 const nickname = () => {
-  const router = useRouter();
-  const [nickName, setNickName] = useState('');
-  const { setName } = useUser();
-
-  const handleClick = () => {
-    router.push('/register/ingredient?type=vegeterian');
-    setName(nickName);
-  }
-
-  const handleNameChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setNickName(event.target.value);
-  };
-
   return (
-    <>
-    <Header title={'NickName'}/>
-    <Box sx={container}>
-    <TextField sx={textFieldStyle} id='outlined-basic' label='NickName' variant='outlined' onChange={handleNameChange}/>
-    </Box>
-    <Progress num={1} onClick={handleClick}/>
-    </>
+    <Nickname/>
   )
 }
 
 export default nickname;
-
-const container = {
-    display: 'flex',
-    margin: '2rem',
-}
-
-const textFieldStyle = {
-  width: '100%',
-  '& .MuiOutlinedInput-root': {
-    '& fieldset': {
-      borderColor: setColor('main'),
-    },
-    '&:hover fieldset': {
-      borderColor: setColor('emphasize'),
-    },
-    '&.Mui-focused fieldset': {
-      borderColor: setColor('main'),
-    },
-  },
-}
